% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_helpers.R
\name{calc_speed}
\alias{calc_speed}
\title{Calculate wind or current speed from u and v component vectors}
\usage{
calc_speed(l)
}
\arguments{
\item{l}{A list comprising a two environmental arrays (consisting of u and v component arrays) for a particular day.}
}
\value{
The function returns an array comprising wind/current speeds for each cell in the original array.
}
\description{
This function calculates wind or current speed from u and v component vector arrays.
}
\details{
In each cell, speed (m/s) is given by Pythagoras' Theorem.
}
\examples{
#### Read example u and v files into a list
# Define the path to the u and v files:
path <- system.file("WeStCOMS_files/",
                    package = "WeStCOMSExploreR", mustWork = TRUE)
path_u <- paste0(path, "uwind_speed")
path_v <- paste0(path, "vwind_speed")
# Define the source files:
source_u <- list.files(path_u, full.names = TRUE)[1]
source_v <- list.files(path_v, full.names = TRUE)[1]
# Define a list comprising u and v arrays with one element for each day:
l <- lapply(list(source_u, source_v), function(source){
              R.matlab::readMat(source)$data
           })
#### Compute wind speed
wind_speed <- calc_speed(l)
hist(wind_speed)

}
\seealso{
This function can be implemented within \code{\link[WeStCOMSExploreR]{compute_field_from_fvcom}}.
}
\author{
Edward Lavender
}
