% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_length}
\alias{check_length}
\title{Check the length of an input to a parent function}
\usage{
check_length(
  arg = deparse(substitute(input)),
  input,
  req_length,
  req_arg = deparse(substitute(req_length))
)
}
\arguments{
\item{arg}{(optional) A character string which defines the argument of a parent function.}

\item{input}{An object.}

\item{req_length}{A number which defines the required length of \code{input}.}

\item{req_arg}{A character which defines the name of the object which defines the required length.}
}
\value{
The function returns a helpful error message for unnamed lists (ignoring empty lists if requested) or the inputted list unchanged.
}
\description{
This function checks that the length of an input to a parent function is correct. If not, the function returns a helpful error message.
}
\author{
Edward Lavender
}
\keyword{internal}
