% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_names}
\alias{check_names}
\title{Check the names of an object contain required names}
\usage{
check_names(
  arg = deparse(substitute(input)),
  input,
  req,
  extract_names = names,
  type = any
)
}
\arguments{
\item{arg}{A character string which defines the argument of the parent function.}

\item{input}{An object for which the names need to be checked.}

\item{req}{A character vector of required names.}

\item{extract_names}{A function which is used to extract names from \code{input}, such as \code{\link[base]{names}} or \code{\link[base]{colnames}}.}

\item{type}{A function which defines the failure criteria. For example, if \code{type = all}, the function will return an error unless all the names in \code{req} are contained within \code{input}. This is the default. If \code{type = any}, the function will return an error only if none of the names in \code{req} are contained within \code{input}.}
}
\value{
If the input fails the check, the function returns a helpful error message. Otherwise, nothing is returned.
}
\description{
This function checks whether required names are contained within an object. If the object does not contain any/all required names (the precise criteria is controlled by the user), the function returns a helpful error message.
}
\author{
Edward Lavender
}
\keyword{internal}
