% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colour_bar.R
\name{colour_bar}
\alias{colour_bar}
\title{Add a colour bar to a plot}
\usage{
colour_bar(
  uniform_values,
  colour_palette_fn,
  cols = NULL,
  legend_labels,
  variable_name,
  variable_name_line,
  cex = 1.8,
  cex.axis = 1.6
)
}
\arguments{
\item{uniform_values}{A numeric, pretty, uniform sequence of values that encompass the range of the variable for which the colour scale is to be plotted. This sequence should be the same as the sequence used to create colours on the plot originally.}

\item{colour_palette_fn}{The function used to provide colour on the main plot.}

\item{cols}{A sequence of colours can be provided instead of the function above.}

\item{legend_labels}{Legend labels for tick marks.}

\item{variable_name}{A variable name to accompany the colour bar.}

\item{variable_name_line}{A numeric value specifying the number of lines the variable_name is away from the axis.}

\item{cex}{The font size of the title, relative to the default.}

\item{cex.axis}{The font size of the legend labels for tick marks.}
}
\description{
This function adds a colour bar to a plot
}
\author{
Edward Lavender
}
\keyword{internal}
