% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_area_boundaries.R
\docType{data}
\name{dat_area_boundaries}
\alias{dat_area_boundaries}
\title{The sample area around Oban}
\format{
A list with 5 elements.
\describe{
  \item{obanxy}{The coordinates (longitude, latitude) of Oban in decimal degrees.}
  \item{adj}{The value by which the coordinates of Oban have been expanded in all four compass directions to define an area around Oban within which sample model outputs are provided.}
  \item{xlim1}{The longitude (decimal degrees) defining the limit of the sample area to the West of Oban.}
  \item{xlim2}{The longitude (decimal degrees) defining the limit of the sample area to the East of Oban.}
  \item{ylim1}{The latitude (decimal degrees) defining the limit of the sample area to the South of Oban.}
  \item{ylim2}{The longitude (decimal degrees) defining the limit of the sample area to the North of Oban.}
}
}
\usage{
dat_area_boundaries
}
\description{
A list containing information on the sample area for which example model outputs are included in the package.
}
\keyword{datasets}
