% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_h.R
\docType{data}
\name{dat_h}
\alias{dat_h}
\title{The depth of each node below mean sea level}
\format{
A data frame with 802 rows and 2 variables:
\describe{
  \item{node_id}{A unique identifier of each node.}
  \item{h}{The depth (m) of each node below mean sea level.}
}
}
\source{
Dmitry Aleynik
}
\usage{
dat_h
}
\description{
A dataset defining the depth of each node below mean sea level for a subset of the WeStCOMS mesh in an area around Oban (see \code{\link[WeStCOMSExploreR]{dat_area_boundaries}}).
}
\keyword{datasets}
