% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_mesh_around_nodes.R
\docType{data}
\name{dat_mesh_around_nodes}
\alias{dat_mesh_around_nodes}
\title{The WeStCOMS mesh around a sample of nodes in the Oban area}
\format{
A SpatialPolygonsDataFrame (see \code{\link[sp]{SpatialPolygonsDataFrame-class}}) with 709 features (i.e. cells).
}
\source{
The WeStCOMS mesh was designed by Dmitry Aleynik.
}
\usage{
dat_mesh_around_nodes
}
\description{
An unstructured grid surrounding nodes (i.e. based on elements) for a subset of the WeStCOMS mesh in an area around Oban (see \code{\link[WeStCOMSExploreR]{dat_area_boundaries}}).
}
\keyword{datasets}
