% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_nodexy.R
\docType{data}
\name{dat_nodexy}
\alias{dat_nodexy}
\title{The coordinates of each node}
\format{
A data frame with 310 rows and 3 variables:
\describe{
  \item{node_id}{A unique identifier of each node.}
  \item{x}{The longitude (decimal degrees) of each node.}
  \item{y}{The latitude (decimal degrees) of each node.}
}
}
\source{
Dmitry Aleynik
}
\usage{
dat_nodexy
}
\description{
A dataset containing the coordinates (longitude, latitude, in decimal degrees) for each node in a subset of the WeStCOMS mesh in an area around Oban (see \code{\link[fvcom.tbx]{dat_area_boundaries}}).
}
\keyword{datasets}
