% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_temp.R
\docType{data}
\name{dat_temp}
\alias{dat_temp}
\title{A sample of surface temperatures around Oban on 2016-03-01.}
\format{A data frame with 802 rows and 2 variables:
\describe{
  \item{ID}{A unique identifier of each node.}
  \item{fvcom}{The temperature (°C) resolved at that node.}
}}
\source{
Model outputs are extracted WeStCOMS outputs computed by Dmitry Aleynik.
}
\usage{
dat_temp
}
\description{
A dataset containing a subset of nodes and the temperatures (°C) resolved at those nodes for layer = 1 on 2016-03-01 (YYYY-MM-DD) at 00:00 hours.
}
\keyword{datasets}
