% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_tidal_elevation.R
\docType{data}
\name{dat_tidal_elevation}
\alias{dat_tidal_elevation}
\title{A sample of tidal elevations around Oban on 2016-03-01.}
\format{
A data frame with 802 rows and 2 variables:
\describe{
  \item{ID}{A unique identifier of each node.}
  \item{fvcom}{The tidal elevation (m) resolved at that node.}
}
}
\source{
Model outputs are extracted WeStCOMS outputs computed by Dmitry Aleynik.
}
\usage{
dat_tidal_elevation
}
\description{
A dataset containing a subset of nodes and the tidal elevation (m) resolved at those nodes on 2016-03-01 (YYYY-MM-DD) at 00:00 hours.
}
\keyword{datasets}
