% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_uwind_speed.R
\docType{data}
\name{dat_uwind_speed}
\alias{dat_uwind_speed}
\title{A sample of eastward wind speeds around Oban on 2016-03-01.}
\format{
A data frame with 1307 rows and 2 variables:
\describe{
  \item{ID}{A unique identifier of each element.}
  \item{fvcom}{The eastward wind speed (\eqn{ms^{-1}}) resolved at that element.}
}
}
\source{
Model output samples are extracted WeStCOMS outputs computed by Dmitry Aleynik.
}
\usage{
dat_uwind_speed
}
\description{
A dataset containing a subset of elements and the u (easting) component of wind velocity (\eqn{ms^{-1}}) resolved at those elements on 2016-03-01 (YYYY-MM-DD) at 00:00 hours.
}
\keyword{datasets}
