% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_name.R
\name{date_name}
\alias{date_name}
\title{Flip between WeStCOMS file names and dates}
\usage{
date_name(x, define = "date_name")
}
\arguments{
\item{x}{A vector of dates or date names to be converted.}

\item{define}{A character that specifes whether 'x' is to be converted from a date to a date_name (define = "date_name") or from a date_name to a date ("date").}
}
\description{
This function converts a date to the 6 digit code that relates that date to a WeStCOMS file name (or vice versa).
}
\examples{
# 1) Extract the WeStCOMS file name as a numeric object from some dates:
date_name(x = as.Date(c("2016-03-01", "2016-03-02")), define = "date_name")

# 2) Define a sequence of dates from WeStCOMS file names:
date_name(x = c(160101, 170101), define = "date")
}
\author{
Edward Lavender
}
