% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.dir.R
\name{define.dir}
\alias{define.dir}
\title{Create folders in which to save WeStCOMS outputs}
\usage{
define.dir(dir = getwd(), vars = "temp")
}
\arguments{
\item{dir}{A string specifying the directory in which you want to create a folder for each environmental variable.}

\item{vars}{A character vector specifying the names of the folders to create in dir. Folders that already exist will not be overwritten; in this case, the function simply returns a warning.}
}
\description{
This function creates a set of folders in the specified working directory in which WeStCOMS outputs (one file for each day) can be stored.
}
\examples{

\dontrun{
# Define folders for several variables in the working directory:
define.dir(dir = getwd(), vars = c("temp", "salinity", "tidal_elevation", "short_wave"))
}

}
\author{
Edward Lavender
}
