% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define.new2dfield.R
\name{define.new2dfield}
\alias{define.new2dfield}
\title{Define new 2 dimensional hydrodynamic fields}
\usage{
define.new2dfield(
  vars,
  nodexy,
  dir,
  date_name,
  cl = NULL,
  pass2varlist = list(NULL)
)
}
\arguments{
\item{vars}{A character vector of variables for which new files should be created. Currently supported options are: (1) 'thermocline_strength'; (2) 'wind_speed'; (3) 'wind_direction'; (4) current_speed; (5) current velocity; (6) sun angle.}

\item{nodexy}{A dataframe containing node ids and coordinates at which sun angle will be calculated. The dataframe should have three columns: node_id, x and y. See \code{?WeStCOMSExploreR::dat_nodexy} for the dataset included in WeStCOMSExploreR as a guide. Note that is is important that the nodexy dataframe should represent the full mesh, rather than a subset of a mesh if you want to treat sun angle in the same way as other WeStCOMS outputs. For example, if you plan to iterate over variables and select a subset of nodes (via indexing) across which to plot the mesh each time, it is essential that the arrays are all of the same dimensions to avoid indexing issues.}

\item{dir}{The directory from which to load WeStCOMS files and in which to save new files. For each of the inputted \code{vars}, the following folders are assumed to exist: "temp", "uwind_speed", "vwind_speed", "uvelocity", "vvelocity" from which model outputs can be loaded.}

\item{date_name}{A vector of date names that correspond to the names of WeStCOMS files and the dates for which you want to create outputs.}

\item{cl}{(optional) A cluster objected created by the parallel package. If supplied, the algorithm is implemented in parallel. Note that the connection with the cluster is stopped within the function.}

\item{pass2varlist}{A list of character vector of names of objects to export to be passed to the \code{varlist} argument of \code{\link[parallel]{clusterExport}}. This is required if you specify some function arguments via objects, rather than directly. (See the use of \code{path} and \code{vars1} in the examples.) These objects must be located in the global environment.}
}
\description{
This function defines new 2 dimensional hydrodynamic fields from WeStCOMS outputs, such as thermocline strength.
}
\details{
Thermocline strength is approximated as the difference in temperature between the surface and the bottom, which is much computationally faster than calculating other metrics (e.g. the standard deviation in temperature across layers). Note that this means that the thermocline_strength option cannot be applied to the data supplied with this package, which is a small subset of the temperature data that does not include the full 10 layers. Speeds and directions are given in m/s and degrees respectively. Direction is the direction of mass flow. Sun angle is calculated using \code{\link[WeStCOMSExploreR]{sun.angle.across.mesh}}.
}
\examples{
#### (1) Implement function on a single processor
# Define directory of the folder containing temperatures:
path <- system.file("WeStCOMS_files/", package = "WeStCOMSExploreR", mustWork = TRUE)
# Define variables for which we'll create new files:
vars1 <- c("wind_speed", "wind_direction", "sun_angle")
# These vars assume that a 'uwind_speed' and 'vwind_speed' folder exist,
# ... from which the necessary data can be loaded
# Examine of path:
list.files(path)
# Use define.dir to create empty folders for these new variables:
define.dir(dir = path, vars = vars1)
# Define new fields
define.new2dfield(vars = vars1,
                  nodexy = WeStCOMSExploreR::dat_nodexy,
                  dir = path,
                  date_name = c("160301", "160302")
                  )
# List files again:
list.files(path, recursive = TRUE)

#### (2) Implement function on multiple cores
# (This is most beneficial for a long vector of date_name),
# but, for demonstration purposes...:
cl <- parallel::makeCluster(2L)
define.new2dfield(vars = vars1,
                  nodexy = WeStCOMSExploreR::dat_nodexy,
                  dir = path,
                  date_name = c("160301", "160302"),
                  cl = cl,
                  pass2varlist = list("path", "vars1"))
# Note that the connection with the cluster is closed internally within the function.

}
\author{
Edward Lavender
}
