% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_dates2load.R
\name{define_dates2load}
\alias{define_dates2load}
\title{Create a dataframe that specified which WeStCOMS files to load}
\usage{
define_dates2load(
  start_date,
  end_date,
  custom_dates = NULL,
  corrupt_dates = NULL
)
}
\arguments{
\item{start_date}{A date object specifying the earliest date for which you have a WeStCOMS file.}

\item{end_date}{A date object specifying the latest date for which you have a WeStCOMS file. Specifying \code{start_date} and \code{end_date} is appropriate if you want to load all dates between a start and end date. If not, see \code{custom_dates}.}

\item{custom_dates}{A vector of dates, for which to create the dataframe. This is an alternative to specifing \code{start_date} and \code{end_date} (see above).}

\item{corrupt_dates}{A vector of dates for which WeStCOMS files are corrupt and should not be included in the dataframe.}
}
\value{
A dataframe with three columns: date (Date object), year (2 digits, a numeric object) and date_name (6 digits, a numeric object).
}
\description{
This function creates a dataframe that can be used to load WeStCOMS files into R or MATLAB.
}
\examples{

# 1) Define a dataframe for a regular sequence of dates between a start and end date:
define_dates2load(start_date = as.Date("2016-03-01"),
                  end_date = as.Date("2016-03-05"),
                  corrupt_dates = NULL)

# 2) Define a dataframe as above, but removing some corrupt dates:
define_dates2load(start_date = as.Date("2016-03-01"),
                  end_date = as.Date("2016-03-05"),
                  corrupt_dates = as.Date(c("2016-03-02", "2016-03-03"))
                  )

# 3) Define a dataframe for a custom series of dates:
define_dates2load(custom_dates = as.Date(c("2016-03-02", "2016-03-03", "2017-01-01")))

}
\author{
Edward Lavender
}
