% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludes.R
\name{exclude_corrupt}
\alias{exclude_corrupt}
\title{Exclude any rows in a dataframe associated with corrupt FVCOM files}
\usage{
exclude_corrupt(dat, corrupt = NULL)
}
\arguments{
\item{dat}{A dataframe (e.g. containing information necessary to extract FVCOM predictions). The only requirement is an integer column named 'date_name' which contains the 6 digit code of FVCOM file names (see \code{\link[WeStCOMSExploreR]{date_name}}).}

\item{corrupt}{A vector of integers which define the 6 digit code of any corrupt FVCOM files.}
}
\value{
A dataframe, as inputted, but in which any rows associated with corrupt FVCOM files have been excluded. If rows have been excluded, the function returns a warning.
}
\description{
This function screens vector of FVCOM file names in a dataframe and removes any rows which refer to the names of known corrupt files (as defined by the user). This is an important check prior to loading multiple FVCOM files into R.
}
\examples{

\dontrun{
exclude_corrupt(data.frame(date_name = c(160301, 160302, 160303)), 160301)
}

}
\seealso{
This check is implemented by \code{\link[WeStCOMSExploreR]{extract}}.
}
\author{
Edward Lavender
}
