% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fvcom.tbx-package.R
\docType{package}
\name{fvcom.tbx}
\alias{fvcom.tbx-package}
\alias{fvcom.tbx}
\title{fvcom.tbx: An R Toolbox for the Unstructured Grid Finite Volume Community Ocean Model (FVCOM)}
\description{
A toolbox for the exploration of Finite Coastal Ocean Volume Model (FVCOM) predictions. Package development has been motivated specifically by the West Coast of Scotland Coastal Modelling System (WeStCOMS). This is a numerical, hydrostatic, hydrodynamic modelling system that can be used to resolve environmental conditions off the West Coast of Scotland, developed at the Scottish Association for Marine Sciences by Dmitry Aleynik. The aim of this package is to provide a gateway for ecological researchers with experience in R into the use of FVCOM predictions for the investigation of spatiotemporal environmental variation in their study sites and the integration of model predictions with ecological research in these environments more widely. See \code{help(package = "fvcom.tbx")} for a full list of package functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/edwardlavender/fvcom.tbx}
  \item Report bugs at \url{http://github.com/edwardlavender/fvcom.tbx/issues}
}

}
\author{
Edward Lavender
}
