% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hours.R
\name{hour_dbl}
\alias{hour_dbl}
\title{Compute the time of day, in hours, as a double}
\usage{
hour_dbl(time)
}
\arguments{
\item{time}{A vector of times, of the class \code{\link[base]{DateTimeClasses}}.}
}
\value{
The function returns a double or vector of doubles which defines/define the time of day, in hours.
}
\description{
Compute the time of day, in hours, from a \code{\link[base]{DateTimeClasses}} object, accounting for the hour of day, the number of minutes passed the hour and the number of seconds passed the minute.
}
\examples{

hour_dbl(as.POSIXct("2016-01-01 13:01:40"))

}
\author{
Edward Lavender
}
