% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hours.R
\name{hour_nearest}
\alias{hour_nearest}
\title{Compute the nearest integer hour of day at supplied times}
\usage{
hour_nearest(time)
}
\arguments{
\item{time}{A vector of times, of the class \code{\link[base]{DateTimeClasses}}.}
}
\value{
The function returns an integer for each inputted time.
}
\description{
Compute the the nearest integer hour of day from a \code{\link[base]{DateTimeClasses}} object (i.e., nearest neighbour interpolation).
}
\examples{

hour_nearest(as.POSIXct("2016-01-01 13:01:40"))
hour_nearest(as.POSIXct(c("2016-01-01 13:01:40", NA)))

}
\author{
Edward Lavender
}
