% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_ts_nearest_by_key}
\alias{match_ts_nearest_by_key}
\title{Match timeseries by key and time}
\usage{
match_ts_nearest_by_key(d1, d2, key_col, time_col)
}
\arguments{
\item{d1}{A dataframe which includes a column that defines factor levels and a column that defines timestamps. The names of these columns need to match those in \code{d2}.}

\item{d2}{A dataframe which includes a column that defines factor levels and a column that defines timestamps. The names of these columns need to match those in \code{d1}.}

\item{key_col}{A character that defines the column name in \code{d1} and \code{d2} that distinguishes factor levels.}

\item{time_col}{A character that defines the column name in \code{d1} and \code{d2} that defines timestamps.}
}
\value{
For a dataframe comprising observations from a series of factor levels (e.g., individuals) collected through time, the function returns a vector of positions in a second dataframe which, for the appropriate factor level, are nearest in time.
}
\description{
For two dataframes, \code{d1} and \code{d2}, this function finds the positions in the second dataframe which, for each key (e.g., factor level, individual) in the first dataframe, are nearest in time (i.e., nearest neighbour interpolation accounting for observations from different factor levels).
}
\details{
If there are multiple matches, only the first is returned.
}
\author{
Edward Lavender
}
\keyword{internal}
