% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quiver.R
\name{quiver}
\alias{quiver}
\title{quiver}
\source{
This function is copied from pracma::quiver but corrects a small bug in the function which stops the angle of arrow heads being adjustable.
}
\usage{
quiver(x, y, u, v, scale = 0.05, angle = 10, length = 0.1, ...)
}
\arguments{
\item{x}{A numeric value specifying the x coordinate(s) of the start point of the arrow(s).}

\item{y}{A numeric value specifying the x coordinate(s) of the start point of the arrow(s).}

\item{u}{The u (Easting) component vector.}

\item{v}{The v (Northing) component vector.}

\item{scale}{A numeric value which scales the length of the arrow shaft.}

\item{angle}{A numeric value which defines the angle (degrees) between the two sides of the arrow head.}

\item{length}{A numeric value which adjusts the length of the arrow head.}

\item{...}{Other graphical parameters that can be passed to \code{\link[graphics]{arrows}}.}
}
\description{
Add a vector field to a plot
}
\note{
This function is not exported.
}
\keyword{internal}
