% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thredds.R
\name{thredds_extract}
\alias{thredds_extract}
\title{Extract FVCOM outputs from the SAMS thredds server}
\usage{
thredds_extract(dat, var, server_catalog, verbose = TRUE, prompt = FALSE)
}
\arguments{
\item{dat}{A dataframe which defines the FVCOM array date names, hours, layers (if applicable) and mesh cell IDs for which FVCOM predictions should be extracted from FVCOM arrays (see \code{\link[fvcom.tbx]{extract}}).}

\item{var}{A character which defines the FVCOM variable for which predictions are to be extracted.}

\item{server_catalog}{A character that defines the URL of the catalog.}

\item{verbose}{A logical input that defines whether or not to print messages to the console. These can be useful for monitoring progress.}

\item{prompt}{A logical input that defines whether or not to prompt the user before each query.}
}
\value{
The function returns a list, with one element for each date name in \code{dat} that contains the associated information in \code{dat} and model predictions.
}
\description{
This function is used to extract FVCOM outputs from the SAMS thredds server.
}
\details{
This function can only be used to query .nc files.

All files corresponding to date names in \code{dat} must be located in \code{server_catalog},

Across all files, only one variable can be queried.

For each file, an iterative approach is used to extract the predictions for hour/layer/mesh IDs specified in \code{dat}. (This may be refined in the future.)
}
\author{
Edward Lavender
}
