% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dir}
\alias{check_dir}
\title{Check a directory exists}
\usage{
check_dir(arg = deparse(substitute(input)), input)
}
\arguments{
\item{arg}{(optional) A character string which defines the argument of a parent function.}

\item{input}{A character string which defines a directory.}
}
\value{
The function checks whether or not a directory exists and, if not, returns an informative error message.
}
\description{
This function checks whether a directory exists and, if not, returns an informative error message.
}
\author{
Edward Lavender
}
\keyword{internal}
