% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_coast_around_oban.R
\docType{data}
\name{dat_coast_around_oban}
\alias{dat_coast_around_oban}
\title{The coastline around Oban}
\format{
A SpatialPolygonsDataFrame (see \code{\link[sp]{SpatialPolygonsDataFrame-class}}).
}
\source{
https://biogeo.ucdavis.edu/data/gadm3.6/Rsp/gadm36_GBR_0_sp.rds
}
\usage{
dat_coast_around_oban
}
\description{
A SpatialPolygonsDataFrame delineating the coastline around Oban in an area defined by \code{\link[fvcom.tbx]{dat_area_boundaries}}.
}
\keyword{datasets}
