% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_mesh_around_elements.R
\docType{data}
\name{dat_mesh_around_elements}
\alias{dat_mesh_around_elements}
\title{The WeStCOMS mesh around a sample of elements in the Oban area}
\format{
A SpatialPolygonsDataFrame with 1307 features (i.e. cells).
}
\source{
The WeStCOMS mesh was designed by Dmitry Aleynik.
}
\usage{
dat_mesh_around_elements
}
\description{
An unstructured grid surrounding elements (i.e. based on nodes) in a subset of the WeStCOMS mesh in an area around Oban defined by area_boundaries_sample.
}
\keyword{datasets}
